refactor : lotto 도메인 패키지 로 이동
