+++ src/test/java/lotto/PositiveNumberTest.java
@@ -3,6 +3,7 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
+import lotto.domain.PositiveNumber;
 import org.junit.jupiter.api.Test;
 
 class PositiveNumberTest {
